Move arithmetic expansion parser to separate module
