+++ yash-syntax/src/parser/lex.rs
@@ -36,8 +36,6 @@ use crate::source::Location;
 use crate::source::SourceChar;
 use crate::syntax::*;
 use std::convert::TryFrom;
-use std::future::Future;
-use std::pin::Pin;
 
 /// Tests whether the given character is a token delimiter.
 ///
@@ -47,74 +45,6 @@ pub fn is_token_delimiter_char(c: char) -> bool {
 }
 
 impl Lexer {
-    /// Parses an arithmetic expansion.
-    ///
-    /// The initial `$` must have been consumed before calling this function.
-    /// In this function, the next two characters are examined to see if they
-    /// begin an arithmetic expansion. If the characters are `((`, then the
-    /// arithmetic expansion is parsed, in which case this function consumes up
-    /// to the closing `))` (inclusive). Otherwise, no characters are consumed
-    /// and the return value is `Ok(Err(opening_location))`.
-    ///
-    /// The `location` parameter should be the location of the initial `$`. It
-    /// is used to construct the result, but this function does not check if it
-    /// actually is a location of `$`.
-    ///
-    /// This function does not consume line continuations between `$` and `(`.
-    /// Line continuations should have been consumed beforehand.
-    pub async fn arithmetic_expansion(
-        &mut self,
-        location: Location,
-    ) -> Result<std::result::Result<TextUnit, Location>> {
-        let index = self.index();
-
-        // Part 1: Parse `((`
-        if !self.skip_if(|c| c == '(').await? {
-            return Ok(Err(location));
-        }
-        self.line_continuations().await?;
-        if !self.skip_if(|c| c == '(').await? {
-            self.rewind(index);
-            return Ok(Err(location));
-        }
-
-        // Part 2: Parse the content
-        let is_delimiter = |c| c == ')';
-        let is_escapable = |c| matches!(c, '$' | '`' | '\\');
-        // Boxing needed for recursion
-        let content: Pin<Box<dyn Future<Output = Result<Text>>>> =
-            Box::pin(self.text_with_parentheses(is_delimiter, is_escapable));
-        let content = content.await?;
-
-        // Part 3: Parse `))`
-        match self.peek_char().await? {
-            Some(sc) if sc.value == ')' => self.consume_char(),
-            Some(_) => unreachable!(),
-            None => {
-                let opening_location = location;
-                let cause = SyntaxError::UnclosedArith { opening_location }.into();
-                let location = self.location().await?.clone();
-                return Err(Error { cause, location });
-            }
-        }
-        self.line_continuations().await?;
-        match self.peek_char().await? {
-            Some(sc) if sc.value == ')' => self.consume_char(),
-            Some(_) => {
-                self.rewind(index);
-                return Ok(Err(location));
-            }
-            None => {
-                let opening_location = location;
-                let cause = SyntaxError::UnclosedArith { opening_location }.into();
-                let location = self.location().await?.clone();
-                return Err(Error { cause, location });
-            }
-        }
-
-        Ok(Ok(TextUnit::Arith { content, location }))
-    }
-
     /// Parses a text unit that starts with `$`.
     ///
     /// If the next character is `$`, a parameter expansion, command
@@ -475,6 +405,7 @@ impl Lexer {
 }
 
 // This is here to get better order of Lexer members in the doc.
+mod arith;
 mod command_subst;
 mod heredoc;
 pub mod keyword;
@@ -485,150 +416,6 @@ mod tests {
     use crate::parser::core::ErrorCause;
     use crate::source::Source;
     use futures::executor::block_on;
-
-    #[test]
-    fn lexer_arithmetic_expansion_empty() {
-        let mut lexer = Lexer::with_source(Source::Unknown, "(());");
-        let location = Location::dummy("X".to_string());
-
-        let result = block_on(lexer.arithmetic_expansion(location))
-            .unwrap()
-            .unwrap();
-        if let TextUnit::Arith { content, location } = result {
-            assert_eq!(content.0, []);
-            assert_eq!(location.line.value, "X");
-            assert_eq!(location.line.number.get(), 1);
-            assert_eq!(location.line.source, Source::Unknown);
-            assert_eq!(location.column.get(), 1);
-        } else {
-            panic!("Not an arithmetic expansion: {:?}", result);
-        }
-
-        assert_eq!(block_on(lexer.peek_char()).unwrap().unwrap().value, ';');
-    }
-
-    #[test]
-    fn lexer_arithmetic_expansion_none() {
-        let mut lexer = Lexer::with_source(Source::Unknown, "( foo bar )baz");
-        let location = Location::dummy("Y".to_string());
-
-        let location = block_on(lexer.arithmetic_expansion(location))
-            .unwrap()
-            .unwrap_err();
-        assert_eq!(location.line.value, "Y");
-        assert_eq!(location.line.number.get(), 1);
-        assert_eq!(location.line.source, Source::Unknown);
-        assert_eq!(location.column.get(), 1);
-
-        assert_eq!(block_on(lexer.peek_char()).unwrap().unwrap().value, '(');
-    }
-
-    #[test]
-    fn lexer_arithmetic_expansion_line_continuations() {
-        let mut lexer = Lexer::with_source(Source::Unknown, "(\\\n\\\n(\\\n)\\\n\\\n);");
-        let location = Location::dummy("X".to_string());
-
-        let result = block_on(lexer.arithmetic_expansion(location))
-            .unwrap()
-            .unwrap();
-        if let TextUnit::Arith { content, location } = result {
-            assert_eq!(content.0, []);
-            assert_eq!(location.line.value, "X");
-            assert_eq!(location.line.number.get(), 1);
-            assert_eq!(location.line.source, Source::Unknown);
-            assert_eq!(location.column.get(), 1);
-        } else {
-            panic!("Not an arithmetic expansion: {:?}", result);
-        }
-
-        assert_eq!(block_on(lexer.peek_char()).unwrap().unwrap().value, ';');
-    }
-
-    #[test]
-    fn lexer_arithmetic_expansion_escapes() {
-        let mut lexer = Lexer::with_source(Source::Unknown, r#"((\\\"\`\$));"#);
-        let location = Location::dummy("X".to_string());
-
-        let result = block_on(lexer.arithmetic_expansion(location))
-            .unwrap()
-            .unwrap();
-        if let TextUnit::Arith { content, location } = result {
-            assert_eq!(
-                content.0,
-                [
-                    Backslashed('\\'),
-                    Literal('\\'),
-                    Literal('"'),
-                    Backslashed('`'),
-                    Backslashed('$')
-                ]
-            );
-            assert_eq!(location.line.value, "X");
-            assert_eq!(location.line.number.get(), 1);
-            assert_eq!(location.line.source, Source::Unknown);
-            assert_eq!(location.column.get(), 1);
-        } else {
-            panic!("Not an arithmetic expansion: {:?}", result);
-        }
-
-        assert_eq!(block_on(lexer.peek_char()).unwrap().unwrap().value, ';');
-    }
-
-    #[test]
-    fn lexer_arithmetic_expansion_unclosed_first() {
-        let mut lexer = Lexer::with_source(Source::Unknown, "((1");
-        let location = Location::dummy("Z".to_string());
-
-        let e = block_on(lexer.arithmetic_expansion(location)).unwrap_err();
-        if let ErrorCause::Syntax(SyntaxError::UnclosedArith { opening_location }) = e.cause {
-            assert_eq!(opening_location.line.value, "Z");
-            assert_eq!(opening_location.line.number.get(), 1);
-            assert_eq!(opening_location.line.source, Source::Unknown);
-            assert_eq!(opening_location.column.get(), 1);
-        } else {
-            panic!("unexpected error cause {:?}", e);
-        }
-        assert_eq!(e.location.line.value, "((1");
-        assert_eq!(e.location.line.number.get(), 1);
-        assert_eq!(e.location.line.source, Source::Unknown);
-        assert_eq!(e.location.column.get(), 4);
-    }
-
-    #[test]
-    fn lexer_arithmetic_expansion_unclosed_second() {
-        let mut lexer = Lexer::with_source(Source::Unknown, "((1)");
-        let location = Location::dummy("Z".to_string());
-
-        let e = block_on(lexer.arithmetic_expansion(location)).unwrap_err();
-        if let ErrorCause::Syntax(SyntaxError::UnclosedArith { opening_location }) = e.cause {
-            assert_eq!(opening_location.line.value, "Z");
-            assert_eq!(opening_location.line.number.get(), 1);
-            assert_eq!(opening_location.line.source, Source::Unknown);
-            assert_eq!(opening_location.column.get(), 1);
-        } else {
-            panic!("unexpected error cause {:?}", e);
-        }
-        assert_eq!(e.location.line.value, "((1)");
-        assert_eq!(e.location.line.number.get(), 1);
-        assert_eq!(e.location.line.source, Source::Unknown);
-        assert_eq!(e.location.column.get(), 5);
-    }
-
-    #[test]
-    fn lexer_arithmetic_expansion_unclosed_but_maybe_command_substitution() {
-        let mut lexer = Lexer::with_source(Source::Unknown, "((1) ");
-        let location = Location::dummy("Z".to_string());
-
-        let location = block_on(lexer.arithmetic_expansion(location))
-            .unwrap()
-            .unwrap_err();
-        assert_eq!(location.line.value, "Z");
-        assert_eq!(location.line.number.get(), 1);
-        assert_eq!(location.line.source, Source::Unknown);
-        assert_eq!(location.column.get(), 1);
-
-        assert_eq!(lexer.index(), 0);
-    }
 
     #[test]
     fn lexer_dollar_unit_no_dollar() {
